feat: implement comment addition functionality and refactor comment handling

+++ back/src/main/java/com/openclassrooms/mddapi/features/comment/CommentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
 
     List<CommentResponseDto> findByPost_IdOrderByCreatedAtAsc(Integer postId);
 }
